# openapi.json
{
  "components": {
    "schemas": {
...
      "CountryCreate": {
        "additionalProperties": false,
        "description": "Request body for creating a trip country.",
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CountryCreate",
        "type": "object"
      },
...
      "CountryOut": {
        "description": "Wire representation of a trip country.",
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "flag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flag"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "item_count": {
            "title": "Item Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "code",
          "flag",
          "is_default",
          "sort_order",
          "item_count"
        ],
        "title": "CountryOut",
        "type": "object"
      },
...
      "CurrencyCreate": {
        "additionalProperties": false,
        "description": "Request body for creating a trip currency.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "is_primary": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Primary"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "required": [
          "code"
        ],
        "title": "CurrencyCreate",
        "type": "object"
      },
...
      "CurrencyOut": {
        "description": "Wire representation of a trip currency.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_primary": {
            "title": "Is Primary",
            "type": "boolean"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "required": [
          "id",
          "code",
          "symbol",
          "is_primary",
          "sort_order"
        ],
        "title": "CurrencyOut",
        "type": "object"
      },
...
      "ErrorBodyOut": {
        "description": "The body of an error: a top-level code, its params, and optional per-field detail.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/FieldErrorOut"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorBodyOut",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "`{ \"error\": { \"code\": ..., \"params\": {...}, \"fields\": {...}? } }`.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyOut"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "FieldErrorOut": {
        "description": "One field's `{code, params}` failure.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code"
        ],
        "title": "FieldErrorOut",
        "type": "object"
      },
...
      "PersonCreate": {
        "additionalProperties": false,
        "description": "Request body for creating a person on a trip.",
        "properties": {
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "default_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Weight"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "PersonCreate",
        "type": "object"
      },
...
      "PersonOut": {
        "description": "Wire representation of a person on a trip.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "color": {
            "title": "Color",
            "type": "string"
          },
          "default_weight": {
            "title": "Default Weight",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "initial": {
            "title": "Initial",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "initial",
          "color",
          "default_weight",
          "sort_order",
          "active"
        ],
        "title": "PersonOut",
        "type": "object"
      },
...
      "TripCreate": {
        "additionalProperties": false,
        "description": "Request body for creating a trip with its roster, currencies, and countries.",
        "properties": {
          "countries": {
            "items": {
              "$ref": "#/components/schemas/CountryCreate"
            },
            "title": "Countries",
            "type": "array"
          },
          "currencies": {
            "items": {
              "$ref": "#/components/schemas/CurrencyCreate"
            },
            "title": "Currencies",
            "type": "array"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "labels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonCreate"
            },
            "title": "People",
            "type": "array"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "required": [
          "name",
          "people",
          "currencies",
          "countries"
        ],
        "title": "TripCreate",
        "type": "object"
      },
      "TripEnvelope": {
        "description": "`{ \"trip\": Trip }`.",
        "properties": {
          "trip": {
            "$ref": "#/components/schemas/TripOut"
          }
        },
        "required": [
          "trip"
        ],
        "title": "TripEnvelope",
        "type": "object"
      },
      "TripListEnvelope": {
        "description": "`{ \"trips\": [TripSummary] }`.",
        "properties": {
          "trips": {
            "items": {
              "$ref": "#/components/schemas/TripSummaryOut"
            },
            "title": "Trips",
            "type": "array"
          }
        },
        "required": [
          "trips"
        ],
        "title": "TripListEnvelope",
        "type": "object"
      },
      "TripOut": {
        "description": "Wire representation of a full trip, including its roster.",
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "countries": {
            "items": {
              "$ref": "#/components/schemas/CountryOut"
            },
            "title": "Countries",
            "type": "array"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "currencies": {
            "items": {
              "$ref": "#/components/schemas/CurrencyOut"
            },
            "title": "Currencies",
            "type": "array"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonOut"
            },
            "title": "People",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "start_date",
          "end_date",
          "note",
          "archived",
          "people",
          "currencies",
          "countries",
          "created_at",
          "updated_at"
        ],
        "title": "TripOut",
        "type": "object"
      },
      "TripSummaryOut": {
        "description": "Wire representation of a trip summary for list views.",
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "item_count": {
            "title": "Item Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "people_count": {
            "title": "People Count",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "start_date",
          "end_date",
          "note",
          "archived",
          "people_count",
          "item_count",
          "created_at",
          "updated_at"
        ],
        "title": "TripSummaryOut",
        "type": "object"
      },
      "TripUpdate": {
        "additionalProperties": false,
        "description": "Request body for partially updating a trip.",
        "properties": {
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "title": "TripUpdate",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Expense tracking for a shared trip: people, currencies, countries and labels belong to a trip; every line item records who paid and how the cost splits across the roster.\n\nJSON in, JSON out. Every response is a single object with its payload under a named key, never a bare array. Amounts are plain numbers in the currency's own units - typed values to 2 fraction digits, computed shares to 6 - and are taken on input as canonical decimal strings (`\"18400.50\"`). Nothing is ever converted between currencies: balances and statistics are per currency.\n\nErrors are `{code, params}` with no message text, so a client renders them through its own catalog in its own language.",
    "title": "Trip Accounter API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/trips": {
      "get": {
        "description": "List all trips, newest first.",
        "operationId": "list_trips_api_v1_trips_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Trips",
        "tags": [
          "trips"
        ]
      },
      "post": {
        "description": "Create a new trip with its people, currencies, and countries.",
        "operationId": "create_trip_api_v1_trips_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{slug}": {
      "delete": {
        "description": "Delete a trip.",
        "operationId": "delete_trip_api_v1_trips__slug__delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Delete Trip",
        "tags": [
          "trips"
        ]
      },
      "get": {
        "description": "Get a single trip by slug.",
        "operationId": "get_trip_route_api_v1_trips__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Trip Route",
        "tags": [
          "trips"
        ]
      },
      "patch": {
        "description": "Update a trip's fields.",
        "operationId": "update_trip_api_v1_trips__slug__patch",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Update Trip",
        "tags": [
          "trips"
        ]
      }
    },
...
  }
}